fixed invoice link in the order view and add logic for deleting files

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const { validationResult } = require("express-validator/check");
 const mongodb = require("mongodb");
+
+const fileHelper = require("../util/file");
 const Product = require("../models/product");
 
 const ObjectId = mongodb.ObjectId;
@@ -181,6 +183,7 @@ exports.postEditProduct = (req, res, next) => {
 			product.price = updatedPrice;
 			product.description = updatedDesc;
 			if (image) {
+				fileHelper.deleteFile(product.imageUrl); //fire and forget manner!means we don't care about the result
 				product.imageUrl = image.path;
 			}
 			return product.save().then((result) => {
@@ -222,9 +225,16 @@ exports.getProducts = (req, res, next) => {
 // destroy() method will delete the entry from the database
 exports.postDeleteProduct = (req, res, next) => {
 	const prodId = req.body.productId;
-	// Product.deleteById(prodId) //mongodb method
-	// Product.findByIdAndRemove(prodId) //mongoose method
-	Product.deleteOne({ _id: prodId, userId: req.user._id })
+	Product.findById(prodId)
+		.then((product) => {
+			if (!product) {
+				return next(new Error("Product not found"));
+			}
+			fileHelper.deleteFile(product.imageUrl); //fire and forget manner!means we don't care about the result
+			return Product.deleteOne({ _id: prodId, userId: req.user._id });
+		})
+		// Product.deleteById(prodId) //mongodb method
+		// Product.findByIdAndRemove(prodId) //mongoose method
 
 		// Product.findByPk(prodId)
 		// 	.then((product) => {
